app/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.10.1
click==8.1.7
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
inflection==0.5.1
isort==5.12.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
PyJWT==1.7.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
sqlparse==0.4.4
uritemplate==4.1.1
